% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssq.R
\name{mssq}
\alias{mssq}
\title{Fit zero-inflated beta regression with random effects}
\usage{
mssq(X, tc, l, N, S, max.diff = 0.001, max.EM.ita = 2000,
  sum.by.markers = NA, species.ind = NA, species.names = NA,
  estimate.phi = TRUE, filter.empty.marker = FALSE, verbose = FALSE)
}
\arguments{
\item{verbose}{print the fitting process}

\item{logistic.cov}{the covariates in logistic component}

\item{beta.cov}{the covariates in beta component}

\item{Y}{the response variable in the regression model}

\item{subject.ind}{the variable for subjects}

\item{time.ind}{the variable for time points}

\item{quad.n}{Gaussian quadrature points}
}
\value{
abu,theta,phi,converge
}
\description{
Fit zero-inflated beta regression with random effects
}
\examples{
\dontrun{
mssq(logistic.cov=logistic.cov,beta.cov=beta.cov,Y=Y,subject.ind=subject.ind,time.ind=time.ind)
}
}

